fix(deeplink): remove context approach for real devices
Using the context approach in `openDeeplinkIosRealDevices` is faster than the method from `openDeeplinkIosSimulators` and depends less on a clean state of the device, but in turn it's very brittle.
Almost every appium update fixes context / WebView issues constantly and still there are cases that are not covered, see for example:
https://github.com/appium/appium/issues/14178
So we switch to the approach in `openDeeplinkIosSimulators` which works universally.