refactor(selector): explicitly mark methods `ios` and `android` as internal and private
BREAKING CHANGE: There is no reason to use the methods `android` and `ios` on `Selector` as part of the public API, which is why they were marked as internal explicitly and renamed to `_android` and `_ios` respectively.
This should make it easier to not use the API in the wrong way.
No migration steps should be necessary for the reasons mentioned above.
If you have previously been using those methods, remember to wrap the `Selector` in either `mobile$` or `mobile$$`.

Signed-off-by: martinfrancois <f.martin@fastmail.com>
(cherry picked from commit a22e116c8577c0bbbc4535a6801d616dc6edf861)